+++ src/services/Element.php
@@ -170,9 +170,9 @@ class Element extends Component
         }
 
         // Populate the record attributes
-        $record->setAttributes(
-            ArrayHelper::toArray($organization)
-        );
+        $record->id = $organization->getId();
+        $record->state = $organization->state;
+        $record->dateJoined = $organization->dateJoined;
 
         return $record;
     }